api/pyproject.toml
[project]
name = "yeshuman-api"
version = "0.1.0"
description = "Yes Human Agent Stack API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ninja>=1.4.3",
    "langgraph>=0.6.5",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.1.0",
    "daphne>=4.2.1",
    "django-ninja-jwt>=5.3.7",
    "django-cors-headers>=4.7.0",
    "psycopg2-binary>=2.9.9",
    "django-polymorphic>=3.1",
    "dj-database-url>=2.1.0",
    "pgvector>=0.4.1",
    "django>=5.2.5",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.0",
    "pytest-watch>=4.2.0",
    "pytest-asyncio>=0.24.0",
    "factory-boy>=3.3.0",
    "faker>=26.0.0",
]
